Minor cosmetic changes and warning fixes
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1810 610dfd11-2700-0410-9b68-9ae28eac5983

+++ include/H3D/HAnimDisplacer.h
@@ -107,7 +107,7 @@ namespace H3D {
 
     for( size_t i = 0; i < indices.size(); i++ ) {
       int index = indices[i];
-      if( index < p_size && i < d_size) {
+      if( (unsigned int) index < p_size && i < d_size) {
         Vec4f point_displacement = displace_transform * Vec4f( displace[i].x, displace[i].y, displace[i].z, 0 ) ;
         points[index] = w * (Vec3f(point_displacement.x,
                                    point_displacement.y,
